requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2021.5.30
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==2.0.4
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==2.2
django-cors-headers==3.7.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
hashids==1.3.1
hyperlink==21.0.0
idna==3.2
incremental==22.10.0
msgpack==1.0.8
Pillow==8.3.1
psycopg2-binary==2.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.1.0
pyOpenSSL==24.1.0
python-decouple==3.4
pytz==2021.1
redis==5.0.7
requests==2.26.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==1.26.6
zope.interface==6.4.post2
